Add segmentation representation and specific configuration
