+++ packages/tools/src/stateManagement/segmentation/helpers/getRepresentationSpecificConfig.ts
@@ -1,0 +1,15 @@
+import SegmentationRepresentations from '../../../enums/SegmentationRepresentations';
+import { RepresentationPublicInput } from '../../../types';
+import { getRepresentationRenderingConfig as getLabelmapRenderingConfig } from '../../../tools/displayTools/Labelmap/labelmapDisplay';
+
+export function getRepresentationSpecificConfig(
+  representationInput: RepresentationPublicInput
+) {
+  const { type } = representationInput;
+
+  if (type === SegmentationRepresentations.Labelmap) {
+    return getLabelmapRenderingConfig();
+  } else {
+    return {};
+  }
+}
