+++ packages/tools/test/segmentationRender_test.js
@@ -17,7 +17,6 @@ const {
   volumeLoader,
   setVolumesForViewports,
   eventTarget,
-  getEnabledElement,
 } = cornerstone3D;
 
 const { unregisterAllImageLoaders } = imageLoader;
